@@ -1,39 +1,50 @@
 import 'package:flutter/material.dart';
 
-void main() => runApp(MaterialApp(
-  home: Home(),
-));
+void main() => runApp(Home());
 
 class Home extends StatelessWidget {
+  final ButtonStyle flatButtonStyle = TextButton.styleFrom(
+    primary: Colors.black87, // Text color
+    backgroundColor: Colors.amber,
+    minimumSize: const Size(88, 36),
+    padding: const EdgeInsets.symmetric(horizontal: 16.0),
+    shape: const RoundedRectangleBorder(
+      borderRadius: BorderRadius.all(Radius.circular(2.0)),
+    ),
+  );
+
+  Home({
+    Key? key,
+  }) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return MaterialApp(
+        home: Scaffold(
       appBar: AppBar(
-          title: Text('my first app'),
+          title: const Text('my first app'),
           centerTitle: true,
-          backgroundColor: Colors.red[600]
-      ),
+          backgroundColor: Colors.red[600]),
       body: Row(
         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
         crossAxisAlignment: CrossAxisAlignment.end,
         children: <Widget>[
-          Text('hello, world'),
-          FlatButton(
+          const Text('hello, world'),
+          TextButton(
+            style: flatButtonStyle,
             onPressed: () {},
-            color: Colors.amber,
-            child: Text('click me'),
+            child: const Text('click me'),
           ),
           Container(
-            color: Colors.cyan,
-            padding: EdgeInsets.all(30.0),
-            child: Text('inside container')
-          ),
+              color: Colors.cyan,
+              padding: const EdgeInsets.all(30.0),
+              child: const Text('inside container')),
         ],
       ),
       floatingActionButton: FloatingActionButton(
-        backgroundColor: Colors.red[600],
-        child: Text('click'),
-      ),
-    );
+          backgroundColor: Colors.red[600],
+          child: const Text('click'),
+          onPressed: () {}),
+    ));
   }
 }
